Seperated and converted to PEP 8 standards
