+++ Pathfinding/GUI/.gitignore
@@ -1,2 +1,3 @@
 .vscode/
 settings.json
+__pycache__
