+++ Pathfinding/GUI/GUI.code-workspace
@@ -9,17 +9,18 @@
     "python.pythonPath": "K:\\Program Files (x86)\\Python37-32\\python.exe",
     "files.exclude": {
       "**/settings.json": true
-      // "*.code-workspace": true
     },
     "editor.wordWrap": "on",
     "editor.formatOnType": true,
     "files.autoSave": "afterDelay",
-    "python.linting.mypyArgs": [
-      "--ignore-missing-imports",
-      "--follow-imports=silent",
-      "--show-column-numbers"
-      // "--extension-pkg-whitelist=pygame",
-      // "--ignored-modules=pygame"
-    ]
+    "python.autoComplete.addBrackets": true,
+
+    "python.linting.flake8Enabled": true,
+    "python.linting.pep8Enabled": true,
+
+    "python.linting.mypyEnabled": false,
+
+    "python.linting.pylintEnabled": true,
+    "python.linting.pylintArgs": ["--extension-pkg-whitelist=pygame"]
   }
 }
